app/requirements.txt
# spec-writer/app/requirements.txt
# ──────────────────────────────────
# Core framework
llama-index==0.12.5

# LLM providers
llama-index-llms-ollama
llama-index-llms-anthropic

# Embedding providers
llama-index-embeddings-ollama
llama-index-embeddings-huggingface

# Vector store
llama-index-vector-stores-chroma
chromadb-client

# Graph store (Phase 3)
llama-index-graph-stores-neo4j

# Code parsing
tree-sitter
tree-sitter-c-sharp
tree-sitter-language-pack

# Structured output
pydantic>=2.0

# API layer (Phase 6)
fastapi
uvicorn[standard]

# Utilities
python-dotenv
rich
nest-asyncio                   